selfcal tests
