+++ tests/test_casa.py
@@ -5,27 +5,6 @@ from __future__ import annotations
 from pathlib import Path
 
 from flint.casa import args_to_casa_task_string
-from flint.selfcal.casa import GainCalOptions, _process_gaincal_options
-
-
-def test_process_gaincal_options() -> None:
-    """Ensure the correct handling of the initialisation of GainCalOptions"""
-
-    empty_gaincal = _process_gaincal_options()
-    assert isinstance(empty_gaincal, GainCalOptions)
-
-    gain_cal_options = GainCalOptions()
-    new_options = _process_gaincal_options(
-        gain_cal_options=gain_cal_options, update_gain_cal_options=None
-    )
-    assert new_options == gain_cal_options
-
-    update_gain_cal_options = {"solint": "inf"}
-    updated_options = _process_gaincal_options(
-        gain_cal_options=gain_cal_options,
-        update_gain_cal_options=update_gain_cal_options,
-    )
-    assert updated_options.solint == "inf"
 
 
 def test_args_to_casa_task_str():
